refactor: made http as server function, removed postFile and putFile functions and moved out classes functions
